[flutter_tools] Add support for custom devices (#78113)

+++ packages/flutter_tools/lib/src/base/config.dart
@@ -28,11 +28,12 @@ class Config {
     required FileSystem fileSystem,
     required Logger logger,
     required Platform platform,
+    bool deleteFileOnFormatException = true
   }) {
     final String filePath = _configPath(platform, fileSystem, name);
     final File file = fileSystem.file(filePath);
     file.parent.createSync(recursive: true);
-    return Config.createForTesting(file, logger);
+    return Config.createForTesting(file, logger, deleteFileOnFormatException: deleteFileOnFormatException);
   }
 
   /// Constructs a new [Config] object from a file called [name] in
@@ -43,14 +44,19 @@ class Config {
     String name = 'test',
     Directory? directory,
     Logger? logger,
+    bool deleteFileOnFormatException = true
   }) {
     directory ??= MemoryFileSystem.test().directory('/');
-    return Config.createForTesting(directory.childFile('.${kConfigDir}_$name'), logger ?? BufferLogger.test());
+    return Config.createForTesting(
+      directory.childFile('.${kConfigDir}_$name'),
+      logger ?? BufferLogger.test(),
+      deleteFileOnFormatException: deleteFileOnFormatException
+    );
   }
 
   /// Test only access to the Config constructor.
   @visibleForTesting
-  Config.createForTesting(File file, Logger logger) : _file = file, _logger = logger {
+  Config.createForTesting(File file, Logger logger, {bool deleteFileOnFormatException = true}) : _file = file, _logger = logger {
     if (!_file.existsSync()) {
       return;
     }
@@ -65,7 +71,10 @@ class Config {
             'You may need to reapply any previously saved configuration '
             'with the "flutter config" command.',
         );
-      _file.deleteSync();
+
+      if (deleteFileOnFormatException) {
+        _file.deleteSync();
+      }
     } on Exception catch (err) {
       _logger
         ..printError('Could not read preferences in ${file.path}.\n$err')
